requirements.txt
langchain
chromadb
sentence-transformers
transformers
faiss-cpu
pdfminer.six
tqdm
pypdf
numpy
requests